Update max length of password
Updated max length of password to 33 characters after reviewing the fantastic detective work Roberto Frenna did in analyzing the payload packet. To summarize: yes, an abnormal packet is created by extending the password array because it overflows into the reserved portion of the packet; however, the abnormal packet still works to unlock telnet, and we need longer password support now that Netgear changed away from using 'Geardog' as the password.